+++ crates/oxc_formatter/src/print/mapped_type.rs
@@ -27,7 +27,9 @@ impl<'a> FormatWrite<'a> for AstNode<'a, TSMappedType<'a>> {
         //   { readonly
         //     [A in B]: T}
         // Because the break is not immediately after `{`.
-        let should_expand = f.source_text().has_newline_after_opening_brace(self.span.start);
+        // `+ 1` skips the opening `{`.
+        let should_expand =
+            f.source_text().has_newline_after_skipping_comments(self.span.start + 1);
 
         let format_inner = format_with(|f| {
             if should_expand {
