+++ crates/oxc_formatter_core/src/spec/suppression.rs
@@ -1,0 +1,5 @@
+/// Returns `true` if a comment body is a suppression marker.
+/// `body` should be the comment text *without* its `//` or `/* */` framing.
+pub fn is_suppression_marker(body: &str) -> bool {
+    matches!(body.trim(), "oxfmt-ignore" | "prettier-ignore")
+}
