+++ crates/oxc_formatter_json/src/lib.rs
@@ -21,5 +21,5 @@ mod separated;
 pub use crate::{
     context::JsonFormatContext,
     format::format,
-    options::{JsonFormatOptions, JsonVariant},
+    options::{BracketSpacing, Expand, JsonFormatOptions, JsonVariant, TrailingCommas},
 };
