+++ crates/oxc_formatter_json/src/print/literal.rs
@@ -4,7 +4,7 @@ use oxc_ast::ast::{NumericLiteral, StringLiteral};
 use oxc_formatter_core::{
     Buffer, Format,
     builders::text,
-    util::{NumberFormatOptions, Quote, format_trimmed_number, normalize_string},
+    spec::{format_trimmed_number, normalize_string},
     write,
 };
 
@@ -29,11 +29,11 @@ impl<'a> Format<'a, JsonFormatContext<'a>> for FmtJsonString<'a, '_> {
             return;
         };
         let inner = &raw_str[1..raw_str.len() - 1];
-        let quotes_will_change = outer_quote != Quote::Double;
+        let quotes_will_change = outer_quote != b'"';
 
         // Prettier's `json` parser normalizes string quotes to `"` and `\r\n` / lone `\r` to `\n`.
         // The shared `normalize_string` handles both.
-        let normalized = normalize_string(inner, Quote::Double, quotes_will_change);
+        let normalized = normalize_string(inner, b'"', quotes_will_change);
 
         // Fast path: outer was already `"` and the body wasn't rewritten.
         if !quotes_will_change && matches!(normalized, Cow::Borrowed(_)) {
@@ -45,15 +45,15 @@ impl<'a> Format<'a, JsonFormatContext<'a>> for FmtJsonString<'a, '_> {
     }
 }
 
-/// Returns the outer quote of `raw`
+/// Returns the outer quote byte of `raw` (`b'"'` / `b'\''`)
 /// if it's a valid quoted-string literal (matching `'…'` or `"…"`),
 /// otherwise `None`.
-fn outer_quote_of(raw: &str) -> Option<Quote> {
+fn outer_quote_of(raw: &str) -> Option<u8> {
     let bytes = raw.as_bytes();
     if bytes.len() < 2 || bytes[0] != bytes[bytes.len() - 1] {
         return None;
     }
-    Quote::from_byte(bytes[0])
+    matches!(bytes[0], b'"' | b'\'').then_some(bytes[0])
 }
 
 pub struct FmtJsonNumber<'a, 'b> {
@@ -66,10 +66,8 @@ impl<'a> Format<'a, JsonFormatContext<'a>> for FmtJsonNumber<'a, '_> {
         // Apply Prettier's number normalization: `.1` → `0.1`, `1.0e+2` → `1.0e2`,
         // `1e00` → `1`, `1.00000` → `1.0`, trailing `.` removal, etc.
         // `keep_one_trailing_decimal_zero` matches Prettier's JS/JSON behavior (`1.00000` → `1.0`, not `1`).
-        let normalized = format_trimmed_number(
-            raw.as_str(),
-            NumberFormatOptions::keep_one_trailing_decimal_zero(),
-        );
+        let normalized =
+            format_trimmed_number(raw.as_str(), /* keep_one_trailing_decimal_zero */ true);
         write!(f, text(arena_cow_str(normalized, f)));
     }
 }
